pyproject.toml
[project]
name = "chessmadness-api"
version = "0.1.0"
description = "FastAPI backend for Chess Repertoire Trainer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "aiosqlite>=0.20",
    "chess>=1.10",
    "pydantic-settings>=2.0",
    "pillow>=11.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9",
    "mypy>=1.14",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
# SQL strings in database.py and routers exceed line-length; ignore E501 there
"app/database.py" = ["E501"]
"app/routers/*.py" = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
